Added request rate limiting and ability to get a specific range of games owned
I added a function that can manage the number of requests that are made to the Steam API per minute. This is disabled by default and can be enabled using a new public function. I also added an alternate function that lets you select a range of games to return for a user's GameList.